+++ src/meltano/api/app.py
@@ -21,6 +21,7 @@ from .workers import MeltanoBackgroundCompiler, UIAvailableWorker, AirflowWorker
 
 
 logger = logging.getLogger(__name__)
+airflow_context = {"worker": None}
 
 
 def create_app(config={}):
@@ -103,6 +104,10 @@ def create_app(config={}):
         init(app)
 
     @app.before_request
+    def setup_airflow_context():
+        g.airflow_worker = airflow_context["worker"]
+
+    @app.before_request
     def setup_js_context():
         appUrl = urlsplit(request.host_url)
         g.jsContext = {"appUrl": appUrl.geturl()[:-1]}
@@ -153,11 +158,10 @@ def start(project, **kwargs):
 
 def start_workers(app, project):
     workers = []
-    try:
-        if not app.config["AIRFLOW_DISABLED"]:
-            workers.append(AirflowWorker(project))
-    except:
-        logger.info("Airflow is not installed.")
+
+    if not app.config["AIRFLOW_DISABLED"]:
+        airflow_context["worker"] = AirflowWorker(project)
+        workers.append(airflow_context["worker"])
 
     workers.append(MeltanoBackgroundCompiler(project))
     workers.append(